Rename new Client from Client2 to Client, replace old one. - Refactored the new client's name from Client2 to Client - Replace the old one - Generate JavaDoc
