--- client/java/doc/index-all.html
@@ -1,203 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc (version 1.7.0_03) on Wed Apr 24 09:36:27 CEST 2013 -->
-<title>Index</title>
-<meta name="date" content="2013-04-24">
-<link rel="stylesheet" type="text/css" href="./stylesheet.css" title="Style">
-</head>
-<body>
-<script type="text/javascript"><!--
-    if (location.href.indexOf('is-external=true') == -1) {
-        parent.document.title="Index";
-    }
-//-->
-</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a name="navbar_top">
-<!--   -->
-</a><a href="#skip-navbar_top" title="Skip navigation links"></a><a name="navbar_top_firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="./package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="./overview-tree.html">Tree</a></li>
-<li><a href="./deprecated-list.html">Deprecated</a></li>
-<li class="navBarCell1Rev">Index</li>
-<li><a href="./help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li>Prev</li>
-<li>Next</li>
-</ul>
-<ul class="navList">
-<li><a href="./index.html?index-all.html" target="_top">Frames</a></li>
-<li><a href="index-all.html" target="_top">No Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_top">
-<li><a href="./allclasses-noframe.html">All Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_top");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip-navbar_top">
-<!--   -->
-</a></div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<div class="contentContainer"><a href="#_A_">A</a>&nbsp;<a href="#_C_">C</a>&nbsp;<a href="#_E_">E</a>&nbsp;<a href="#_G_">G</a>&nbsp;<a href="#_O_">O</a>&nbsp;<a href="#_R_">R</a>&nbsp;<a href="#_S_">S</a>&nbsp;<a name="_A_">
-<!--   -->
-</a>
-<h2 class="title">A</h2>
-<dl>
-<dt><span class="strong"><a href="./Client.html#accept(java.lang.String, char)">accept(String, char)</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Instruct the server to accept a game.</div>
-</dd>
-</dl>
-<a name="_C_">
-<!--   -->
-</a>
-<h2 class="title">C</h2>
-<dl>
-<dt><a href="./Client.html" title="class in &lt;Unnamed&gt;"><span class="strong">Client</span></a> - Class in <a href="./package-summary.html">&lt;Unnamed&gt;</a></dt>
-<dd>
-<div class="block">Provides an interface to the Internet MiniChess Server.</div>
-</dd>
-<dt><span class="strong"><a href="./Client.html#Client(java.lang.String, java.lang.String, java.lang.String, java.lang.String)">Client(String, String, String, String)</a></span> - Constructor for class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Create a new client connected to IMCS and logged on.</div>
-</dd>
-<dt><span class="strong"><a href="./Client.html#Client(java.lang.String, java.lang.String, java.lang.String, java.lang.String, java.lang.String)">Client(String, String, String, String, String)</a></span> - Constructor for class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Create a new client connected to IMCS and logged on.</div>
-</dd>
-<dt><span class="strong"><a href="./Client.html#close()">close()</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Closes the connection to the server.</div>
-</dd>
-</dl>
-<a name="_E_">
-<!--   -->
-</a>
-<h2 class="title">E</h2>
-<dl>
-<dt><span class="strong"><a href="./Client.html#expect(java.lang.String, boolean)">expect(String, boolean)</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Get a response from the server.</div>
-</dd>
-<dt><span class="strong"><a href="./Client.html#expectResponse(boolean)">expectResponse(boolean)</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Get a response from the server.</div>
-</dd>
-</dl>
-<a name="_G_">
-<!--   -->
-</a>
-<h2 class="title">G</h2>
-<dl>
-<dt><span class="strong"><a href="./Client.html#getMove()">getMove()</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Get a move string from the IMCS server.</div>
-</dd>
-</dl>
-<a name="_O_">
-<!--   -->
-</a>
-<h2 class="title">O</h2>
-<dl>
-<dt><span class="strong"><a href="./Client.html#offer(char)">offer(char)</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Instruct the server to offer a game to
- other players.</div>
-</dd>
-</dl>
-<a name="_R_">
-<!--   -->
-</a>
-<h2 class="title">R</h2>
-<dl>
-<dt><span class="strong"><a href="./Client.html#responseCode(java.lang.String)">responseCode(String)</a></span> - Static method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Get the response code from a response.</div>
-</dd>
-<dt><span class="strong"><a href="./Client.html#responseString(java.lang.String)">responseString(String)</a></span> - Static method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Get the response text from a response.</div>
-</dd>
-</dl>
-<a name="_S_">
-<!--   -->
-</a>
-<h2 class="title">S</h2>
-<dl>
-<dt><span class="strong"><a href="./Client.html#send(java.lang.String, boolean)">send(String, boolean)</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Send a command to the server.</div>
-</dd>
-<dt><span class="strong"><a href="./Client.html#sendMove(java.lang.String)">sendMove(String)</a></span> - Method in class <a href="./Client.html" title="class in &lt;Unnamed&gt;">Client</a></dt>
-<dd>
-<div class="block">Send a move to the server.</div>
-</dd>
-</dl>
-<a href="#_A_">A</a>&nbsp;<a href="#_C_">C</a>&nbsp;<a href="#_E_">E</a>&nbsp;<a href="#_G_">G</a>&nbsp;<a href="#_O_">O</a>&nbsp;<a href="#_R_">R</a>&nbsp;<a href="#_S_">S</a>&nbsp;</div>
-<!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a name="navbar_bottom">
-<!--   -->
-</a><a href="#skip-navbar_bottom" title="Skip navigation links"></a><a name="navbar_bottom_firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="./package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="./overview-tree.html">Tree</a></li>
-<li><a href="./deprecated-list.html">Deprecated</a></li>
-<li class="navBarCell1Rev">Index</li>
-<li><a href="./help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li>Prev</li>
-<li>Next</li>
-</ul>
-<ul class="navList">
-<li><a href="./index.html?index-all.html" target="_top">Frames</a></li>
-<li><a href="index-all.html" target="_top">No Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_bottom">
-<li><a href="./allclasses-noframe.html">All Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_bottom");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip-navbar_bottom">
-<!--   -->
-</a></div>
-<!-- ======== END OF BOTTOM NAVBAR ======= -->
-</body>
-</html>
