+++ client/java/docs/allclasses-frame.html
@@ -2,16 +2,18 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (version 1.7.0_03) on Wed Apr 24 09:36:27 CEST 2013 -->
+<!-- Generated by javadoc (1.8.0_121) on Wed May 17 07:47:13 CEST 2017 -->
 <title>All Classes</title>
-<meta name="date" content="2013-04-24">
+<meta name="date" content="2017-05-17">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
+<script type="text/javascript" src="script.js"></script>
 </head>
 <body>
-<h1 class="bar">All Classes</h1>
+<h1 class="bar">All&nbsp;Classes</h1>
 <div class="indexContainer">
 <ul>
 <li><a href="Client.html" title="class in &lt;Unnamed&gt;" target="classFrame">Client</a></li>
+<li><a href="IMCSGame.html" title="class in &lt;Unnamed&gt;" target="classFrame">IMCSGame</a></li>
 </ul>
 </div>
 </body>
