+++ client/java/docs/package-frame.html
@@ -2,10 +2,11 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (version 1.7.0_03) on Wed Apr 24 09:36:27 CEST 2013 -->
-<title>&amp;lt;Unnamed&amp;gt;</title>
-<meta name="date" content="2013-04-24">
+<!-- Generated by javadoc (1.8.0_121) on Wed May 17 07:47:13 CEST 2017 -->
+<title>&lt;Unnamed&gt;</title>
+<meta name="date" content="2017-05-17">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
+<script type="text/javascript" src="script.js"></script>
 </head>
 <body>
 <h1 class="bar"><a href="package-summary.html" target="classFrame">&lt;Unnamed&gt;</a></h1>
@@ -13,6 +14,7 @@
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
 <li><a href="Client.html" title="class in &lt;Unnamed&gt;" target="classFrame">Client</a></li>
+<li><a href="IMCSGame.html" title="class in &lt;Unnamed&gt;" target="classFrame">IMCSGame</a></li>
 </ul>
 </div>
 </body>
