+++ client/java/docs/package-tree.html
@@ -2,24 +2,38 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (version 1.7.0_03) on Wed Apr 24 09:36:27 CEST 2013 -->
-<meta name="date" content="2013-04-24">
+<!-- Generated by javadoc (1.8.0_121) on Wed May 17 07:47:13 CEST 2017 -->
+<title> Class Hierarchy</title>
+<meta name="date" content="2017-05-17">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
+<script type="text/javascript" src="script.js"></script>
 </head>
 <body>
+<script type="text/javascript"><!--
+    try {
+        if (location.href.indexOf('is-external=true') == -1) {
+            parent.document.title=" Class Hierarchy";
+        }
+    }
+    catch(err) {
+    }
+//-->
+</script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
 </noscript>
 <!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a name="navbar_top">
+<div class="topNav"><a name="navbar.top">
 <!--   -->
-</a><a href="#skip-navbar_top" title="Skip navigation links"></a><a name="navbar_top_firstrow">
+</a>
+<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.top.firstrow">
 <!--   -->
 </a>
 <ul class="navList" title="Navigation">
 <li><a href="package-summary.html">Package</a></li>
 <li>Class</li>
-<li><a href="package-tree.html">Tree</a></li>
+<li class="navBarCell1Rev">Tree</li>
 <li><a href="deprecated-list.html">Deprecated</a></li>
 <li><a href="index-all.html">Index</a></li>
 <li><a href="help-doc.html">Help</a></li>
@@ -27,15 +41,15 @@
 </div>
 <div class="subNav">
 <ul class="navList">
-<li>Prev Package</li>
-<li>Next Package</li>
+<li>Prev</li>
+<li>Next</li>
 </ul>
 <ul class="navList">
-<li><a href="index.html?package-summary.html" target="_top">Frames</a></li>
-<li><a href="package-summary.html" target="_top">No Frames</a></li>
+<li><a href="index.html?package-tree.html" target="_top">Frames</a></li>
+<li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
-<li><a href="allclasses-noframe.html">All Classes</a></li>
+<li><a href="allclasses-noframe.html">All&nbsp;Classes</a></li>
 </ul>
 <div>
 <script type="text/javascript"><!--
@@ -49,44 +63,36 @@
   //-->
 </script>
 </div>
-<a name="skip-navbar_top">
+<a name="skip.navbar.top">
 <!--   -->
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;&lt;Unnamed&gt;</h1>
+<h1 class="title">Hierarchy For Package &lt;Unnamed&gt;</h1>
 </div>
 <div class="contentContainer">
-<ul class="blockList">
-<li class="blockList">
-<table class="packageSummary" border="0" cellpadding="3" cellspacing="0" summary="Class Summary table, listing classes, and an explanation">
-<caption><span>Class Summary</span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colFirst" scope="col">Class</th>
-<th class="colLast" scope="col">Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<td class="colFirst"><a href="Client.html" title="class in &lt;Unnamed&gt;">Client</a></td>
-<td class="colLast">
-<div class="block">Provides an interface to the Internet MiniChess Server.</div>
-</td>
-</tr>
-</tbody>
-</table>
+<h2 title="Class Hierarchy">Class Hierarchy</h2>
+<ul>
+<li type="circle">java.lang.Object
+<ul>
+<li type="circle"><a href="Client.html" title="class in &lt;Unnamed&gt;"><span class="typeNameLink">Client</span></a></li>
+<li type="circle"><a href="IMCSGame.html" title="class in &lt;Unnamed&gt;"><span class="typeNameLink">IMCSGame</span></a></li>
+</ul>
 </li>
 </ul>
 </div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a name="navbar_bottom">
+<div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
-</a><a href="#skip-navbar_bottom" title="Skip navigation links"></a><a name="navbar_bottom_firstrow">
+</a>
+<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.bottom.firstrow">
 <!--   -->
 </a>
 <ul class="navList" title="Navigation">
 <li><a href="package-summary.html">Package</a></li>
 <li>Class</li>
-<li><a href="package-tree.html">Tree</a></li>
+<li class="navBarCell1Rev">Tree</li>
 <li><a href="deprecated-list.html">Deprecated</a></li>
 <li><a href="index-all.html">Index</a></li>
 <li><a href="help-doc.html">Help</a></li>
@@ -94,15 +100,15 @@
 </div>
 <div class="subNav">
 <ul class="navList">
-<li>Prev Package</li>
-<li>Next Package</li>
+<li>Prev</li>
+<li>Next</li>
 </ul>
 <ul class="navList">
-<li><a href="index.html?package-summary.html" target="_top">Frames</a></li>
-<li><a href="package-summary.html" target="_top">No Frames</a></li>
+<li><a href="index.html?package-tree.html" target="_top">Frames</a></li>
+<li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
-<li><a href="allclasses-noframe.html">All Classes</a></li>
+<li><a href="allclasses-noframe.html">All&nbsp;Classes</a></li>
 </ul>
 <div>
 <script type="text/javascript"><!--
@@ -116,7 +122,7 @@
   //-->
 </script>
 </div>
-<a name="skip-navbar_bottom">
+<a name="skip.navbar.bottom">
 <!--   -->
 </a></div>
 <!-- ======== END OF BOTTOM NAVBAR ======= -->
